table/policy: Enable to replace whole of the statement
Currently, ReplaceStatement() does not whole of confitions and actions.

For example, assume there is a statement which has conditions
"Rpki(invalid) and AsPathLength(=2)",
and call ReplaceStatement() only with "Rpki(not-found)".
The result is "Rpki(not-found) and AsPathLength(=2)",
because ReplaceStatement() only cares about Rpki, and leaves AsPathLength as it is.

For convinience, it is desireble that ReplaceStatement()
can replace whole of the statement.
This adds an argument to specify whether replace whole of the statement or not.

Signed-off-by: Satoshi Fujimoto <satoshi.fujimoto7@gmail.com>